requirements.txt
pyrogram==2.0.106
tgcrypto
py-tgcalls>=2.2.0,<3.0.0
yt-dlp>=2023.07.06
python-dotenv>=1.0.0
aiohttp>=3.8.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
ffmpeg-python>=0.2.0
pillow>=9.0.0
requests>=2.31.0
